# senai_crm/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SenAI CRM",
    "description": "AI-powered email operations and triage system",
    "version": "0.4.0"
  },
  "paths": {
...
    "/agent/run/{email_id}": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Run the agent on an email (full execution \u2014 writes actions to DB)",
        "operationId": "run_agent_agent_run__email_id__post",
        "parameters": [
          {
            "name": "email_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Email Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agent/dry-run/{email_id}": {
      "post": {
        "tags": [
          "agent"
        ],
        "summary": "Dry-run the agent \u2014 returns full reasoning trace, no DB writes",
        "operationId": "dry_run_agent_agent_dry_run__email_id__post",
        "parameters": [
          {
            "name": "email_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Email Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRunResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/drafts/{draft_id}": {
      "patch": {
        "tags": [
          "drafts"
        ],
        "summary": "Edit a draft's content",
        "operationId": "update_draft_drafts__draft_id__patch",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Draft Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/drafts/{draft_id}/approve": {
      "post": {
        "tags": [
          "drafts"
        ],
        "summary": "Approve a draft action",
        "operationId": "approve_draft_drafts__draft_id__approve_post",
        "parameters": [
          {
            "name": "draft_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Draft Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/contacts/{contact_email}": {
      "get": {
        "tags": [
          "contacts"
        ],
        "summary": "Get contact profile",
        "operationId": "get_contact_contacts__contact_email__get",
        "parameters": [
          {
            "name": "contact_email",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Contact Email"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/contacts/{contact_email}/status": {
      "patch": {
        "tags": [
          "contacts"
        ],
        "summary": "Update contact status",
        "operationId": "update_contact_status_contacts__contact_email__status_patch",
        "parameters": [
          {
            "name": "contact_email",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Contact Email"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContactStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AgentRunResult": {
        "properties": {
          "email_id": {
            "type": "integer",
            "title": "Email Id"
          },
          "message_id": {
            "type": "string",
            "title": "Message Id"
          },
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/AgentStep"
            },
            "type": "array",
            "title": "Steps"
          },
          "final_action": {
            "type": "string",
            "title": "Final Action"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "action_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action Id"
          }
        },
        "type": "object",
        "required": [
          "email_id",
          "message_id",
          "dry_run",
          "steps",
          "final_action",
          "summary"
        ],
        "title": "AgentRunResult"
      },
      "AgentStep": {
        "properties": {
          "step": {
            "type": "integer",
            "title": "Step"
          },
          "thought": {
            "type": "string",
            "title": "Thought"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "action_input": {
            "additionalProperties": true,
            "type": "object",
            "title": "Action Input"
          },
          "observation": {
            "type": "string",
            "title": "Observation"
          }
        },
        "type": "object",
        "required": [
          "step",
          "thought",
          "action",
          "action_input",
          "observation"
        ],
        "title": "AgentStep"
      },
      "ApproveRequest": {
        "properties": {
          "approved_by": {
            "type": "string",
            "title": "Approved By",
            "default": "human"
          }
        },
        "type": "object",
        "title": "ApproveRequest"
      },
...
      "ContactResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "account_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Value"
          },
          "churn_risk_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Churn Risk Score"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "last_contact_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Contact At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "name",
          "company",
          "status",
          "account_value",
          "churn_risk_score",
          "created_at",
          "last_contact_at"
        ],
        "title": "ContactResponse"
      },
      "ContactStatusUpdate": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "performed_by": {
            "type": "string",
            "title": "Performed By",
            "default": "human"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "ContactStatusUpdate"
      },
...
      "DraftResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "email_id": {
            "type": "integer",
            "title": "Email Id"
          },
          "action_type": {
            "type": "string",
            "title": "Action Type"
          },
          "proposed_content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposed Content"
          },
          "is_approved": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Approved"
          },
          "approved_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved By"
          },
          "executed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Executed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email_id",
          "action_type",
          "proposed_content",
          "is_approved",
          "approved_by",
          "executed_at"
        ],
        "title": "DraftResponse"
      },
      "DraftUpdateRequest": {
        "properties": {
          "proposed_content": {
            "type": "string",
            "title": "Proposed Content"
          }
        },
        "type": "object",
        "required": [
          "proposed_content"
        ],
        "title": "DraftUpdateRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}